Use whereami.c to find the path of the irace.exe and ablation.exe executables.
The command-line executables `irace` and `ablation` (`irace.exe` and
`ablation.exe` in Windows) will load the version of the `irace` package that is
found in the same path where the executables are. In earlier versions, the
executables will always load the version found via `base::.libPaths()`.  This
change allows installing multiple versions of the irace package in different
locations and each executable will use its corresponding version.  The correct
location can be verified by looking at the line `"installed at:"` printed in
the output.